+++ app/locale/loader.js
@@ -314,8 +314,13 @@ function translateUI(lang, locale, template) {
   $('#lbl-overlayPosition').text(clear(template.settings.notification.option.overlayPosition));
   $('#lbl-overlayScale').text(clear(template.settings.notification.option.overlayScale));
   $('#lbl-overlaySound').text(clear(template.settings.notification.option.overlaySound));
+  $('#lbl-overlayRandomSound').text(clear(template.settings.notification.option.overlayRandomSound));
+  $('#lbl-overlayPlaytimeScale').text(clear(template.settings.notification.option.overlayPlaytimeScale));
   $('#lbl-overlayVolume').text(clear(template.settings.notification.option.overlayVolume));
   $('#lbl-overlayDuration').text(clear(template.settings.notification.option.overlayDuration));
+  $('#option_overlayRandomSound option[value="false"]').text(clear(template.settings.common.disable));
+  $('#option_overlayRandomSound option[value="true"]').text(clear(template.settings.common.enable));
+  $('#option_overlayPlaytimeScale option[value="0"]').text(clear(template.settings.notification.option.playtimeScaleSameAsMain));
   if (template.settings.notification.option.overlaySoundImport) {
     $('#btn-import-sound').attr('title', clear(template.settings.notification.option.overlaySoundImport));
   }
@@ -324,7 +329,9 @@ function translateUI(lang, locale, template) {
   $('#lbl-overlayPreset').closest('li').find('.help').text(clear(template.settings.notification.option.overlayPresetDesc));
   $('#lbl-overlayPosition').closest('li').find('.help').text(clear(template.settings.notification.option.overlayPositionDesc));
   $('#lbl-overlaySound').closest('li').find('.help').text(clear(template.settings.notification.option.overlaySoundDesc));
+  $('#lbl-overlayRandomSound').closest('li').find('.help').text(clear(template.settings.notification.option.overlayRandomSoundDesc));
   $('#lbl-overlayScale').closest('li').find('.help').text(clear(template.settings.notification.option.overlayScaleDesc));
+  $('#lbl-overlayPlaytimeScale').closest('li').find('.help').text(clear(template.settings.notification.option.overlayPlaytimeScaleDesc));
   $('#lbl-overlayVolume').closest('li').find('.help').text(clear(template.settings.notification.option.overlayVolumeDesc));
   $('#lbl-overlayDuration').closest('li').find('.help').text(clear(template.settings.notification.option.overlayDurationDesc));
   $("#option_overlayDuration option[value='auto']").text(clear(template.settings.notification.option.overlayDurationAuto));